Add files via upload
The is Ben Eater's version of 8-bit programmable breadboard computer in simulation. The simulation is done using proteus 8.10. All files should be downloaded and imported in the EPROM. To do so, go to EPROM property and import the hex file. The hex file for control unit is written to conduct LDA 14, ADD 15 and OUT operation. I will upload simulation containing more instructions and operations after some time.